+++ Models/User/ReserveCase.hpp
@@ -16,20 +16,20 @@ public:
     int get_reserve_id();
     std::string get_hotel_reserved_id();
     int get_quantity_of_reserves();
-    float get_cost_of_reserves();
+    long double get_cost_of_reserves();
     int get_residency_period();
     std::pair<int,int> get_reserved_days();
     std::string get_type_of_reserved_rooms();
 
 private:
-    ReserveCase(Database::Hotel* hotel, v_rooms& b_rooms, float price, int check_in, int check_out, std::list<int>& l_id);
+    ReserveCase(Database::Hotel* hotel, v_rooms& b_rooms, long double price, int check_in, int check_out, std::list<int>& l_id);
     int generate_id(std::list<int>& l_id);
 
 private:
     int id;
     Database::Hotel* hotel;
     v_rooms rooms;
-    float cost;
+    long double cost;
     int check_in;
     int check_out;
 };
